Default handling of nulls added to display the volume name
Purge local storage added to developer options and purges before a search is begun to handle issues with over exceeding the side of local storage.
(Purge not removes all search results however keeps a store of all items in a collection.)

side menu updated in all html to include hyper links and to ensure everything is loaded into the dom correctly.